Updated Master

+++ js/jerkScript.js
@@ -7,7 +7,7 @@ jerkScript.js
 
 /*Game mode: 
 0 = Marathon
-1 = Untimed
+1 = Sudden Death
 2 = Time Attack
 */
 var gamemode;
@@ -42,9 +42,6 @@ var localSavedFiles;
 var colourArray = ["cyan","orange","green","pink","blue","purple"];
 var colour;
 
-/*Keeps track of time in Marathon and Time Attack modes*/
-var counter;
-
 $.getScript("js/gameTimer.js", function(){});
 $.getScript("js/nuggetScript.js", function(){});
 $.getScript("js/pathGenerator.js", function(){});
@@ -70,11 +67,15 @@ function gameSetup() {
             gameSetup();
 		}
 		console.log(localSavedFiles);
-		updateBadges();
 	}
+	updateBadges();
 	updateHighScores();
 	document.getElementById('local-scores').style.display = 'block';
 	document.getElementById('online-scores').style.display = 'none';
+	$('input[type=checkbox]').each(function() { 
+        this.checked = true; 
+	}); 
+	$('#option-4').removeAttr('checked');
 	$("#title").text('DotDash._');
 	$("#main-screen").css("display", "block");
     /*see touching.js */
@@ -95,9 +96,14 @@ function checkCookie(){
 }
 
 function clearSave() {
-	localSavedFiles = null;
-	localStorage.setItem("saveFile", JSON.stringify(localSavedFiles));
-	gameSetup();
+	var resetYes = window.confirm("Are you sure you want to reset your save? Your high scores and badges will be lost.");
+	if (resetYes) {
+		localSavedFiles = [false, false, false, false, false, false, false, false, false, false, false, 0, 0, 0];
+		localStorage.setItem("saveFile", JSON.stringify(localSavedFiles));
+		updateBadges();
+		updateHighScores();	
+		window.alert("Save Cleared");
+	}
 }
 
 function initialize(gamemode, newRound, removeDots){
@@ -114,16 +120,17 @@ function initialize(gamemode, newRound, removeDots){
 	numCols = 3;
 	playing = true;
 	userInput = false;
-    updateLives();
-    updateScore();
 	if (gamemode == 1) {
 		document.getElementById('timer-bar').style.visibility='hidden';
+		lifePoints = 1;
 	} else {
 		document.getElementById('timer-bar').style.visibility='visible'; 
 	}
+	updateLives();
+    updateScore();
 	switch(gamemode) {
 		case 0:
-			timerSet(0, 5, 0);
+			timerSet(0, 10, 0);
 			break;
 		case 1: 
 			timerSet(0, 0, 0);
@@ -164,7 +171,7 @@ function newRound(generateGrid){
 	$("#timer-bar").css("background-color", "#32CD32");
 	if (playing) {
 		if (gamemode == 0) {
-			timerSet(0, 5, 0);
+			timerSet(0, 10, 0);
 		}
 		noErrorsYet = true;
 		notComplete = true;
@@ -211,7 +218,7 @@ function generateGrid(createGrid, make_2D_Array, pathDemonstration){
         }
     }
    // var someGrid = new Grid(numRows,numCols);
-    var arrayToRepeat = runPathFinder(numRows, numCols, difficulty(numRows * numCols), true);
+    var arrayToRepeat = runPathFinder(numRows, numCols, difficulty(numRows * numCols));
     if(steveModeEnabled){
         steveify();
     }
@@ -294,7 +301,7 @@ function difficulty(nodeCount) {
 	} else {
 		numRows = 4;
 		numCols = 4;
-		length = 3 + ((0.02 * currentRound) - 0.02);
+		length = 3 + ((0.11 * currentRound) - 0.11);
 	}
     if (length > nodeCount) {
         return nodeCount;
@@ -313,22 +320,6 @@ function difficulty(nodeCount) {
 function validate(array, userFeedback, dArray){
     var ex, wai;
 	fadeOff();
-	counter = setTimeout(function() {
-		if (gamemode != 1) {
-			console.log("TIMESUP");
-			if (gamemode == 2) {
-				lifePoints = 0;
-				updateLives();
-				return;
-			}
-			noErrorsYet = false;
-			lifePoints--;
-			popupSound.play();
-			updateLives();
-			userFeedback(false, null);
-			return;
-		}
-	}, ((minutes * 60) * 1020) + (seconds * 1020));
 	
     $(function(){
         $( ".dot" ).bind( "tapone", tapHandler );
@@ -362,8 +353,8 @@ function validate(array, userFeedback, dArray){
                         index++;
 						//Player successfully cleared round
                         if (index >= array.length) {
+							userInput = false;
 							levelPass.play();
-							clearTimeout(counter);
 							timerPause();
                             notComplete = false;
                             playerScore += currentRound + (Math.round(currentRound * 0.1) * seconds);
@@ -374,7 +365,7 @@ function validate(array, userFeedback, dArray){
                     } 
 				//Player goofed
                 } else {
-					clearTimeout(counter);
+					userInput = false;
 					timerPause();
 					noErrorsYet = false;
 					if (gamemode == 2) {
@@ -436,7 +427,6 @@ function userFeedback(bool, lastNode) {
 		
         //adjustStats(reset);
 		reset(removeDots);
-
     }, 800);
 }
 
@@ -459,7 +449,6 @@ function removeDots(){
 * Briefly changes the colour of each to indicate which dots should be
  * selected in which sequence.*/
 function pathDemonstration(arrayToRepeat, validate) {
-	userInput = false;
     var pt;
 	var blinkTime;
 	if (currentRound >= 250) {
@@ -514,20 +503,6 @@ function playAgain() {
 }
 
 function resumeGame() {
-	counter = setTimeout(function() {
-		if (gamemode != 1) {
-			if (gamemode == 2) {
-				lifePoints = 0;
-				updateLives();
-				return;
-			}
-			noErrorsYet = false;
-			lifePoints--;
-			updateLives();
-			userFeedback(false, null);
-			return;
-		}
-	}, ((minutes * 60) * 1000) + (seconds * 1000) + 20);
 	if (gamemode != 1) {
 		timerStart();
 	}
@@ -535,7 +510,6 @@ function resumeGame() {
 
 function pauseGame(type) {
     if (userInput) {
-        clearTimeout(counter);
         timerPause();
         if (type == 'pause') {
             openPauseScreen();
@@ -551,7 +525,6 @@ function gameOver() {
 
     document.getElementById('tutorial1-screen').style.display = "none";
     document.getElementById('tutorial4-screen').style.display = "none";
-
     $( "#game-screen" ).fadeOut( 1500, function() {
         $('#gameover-screen').fadeIn(1500, function() {});
 		badgeChecker(currentRound, lifePoints);
